requirements.txt
django==3.2.13
django-model-utils==4.2.0
djangorestframework==3.13.1